fix : MemberRequest -> MemberController
